+++ op-signer/client/client.go
@@ -10,8 +10,8 @@ import (
 	"os"
 	"time"
 
-	"github.com/dyson/certman"
 	optls "github.com/ethereum-optimism/optimism/op-service/tls"
+	"github.com/ethereum-optimism/optimism/op-service/tls/certman"
 	"github.com/ethereum/go-ethereum/common/hexutil"
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/log"
@@ -24,42 +24,44 @@ type SignerClient struct {
 	logger log.Logger
 }
 
-// ethLogger wraps a geth style logger for certman.
-type ethLogger struct{ logger log.Logger }
-
-func (l ethLogger) Printf(format string, v ...interface{}) { l.logger.Info(fmt.Sprintf(format, v...)) }
-
 func NewSignerClient(logger log.Logger, endpoint string, tlsConfig optls.CLIConfig) (*SignerClient, error) {
-	caCert, err := os.ReadFile(tlsConfig.TLSCaCert)
-	if err != nil {
-		return nil, fmt.Errorf("failed to read tls.ca: %w", err)
-	}
-	caCertPool := x509.NewCertPool()
-	caCertPool.AppendCertsFromPEM(caCert)
+	var httpClient *http.Client
+	if tlsConfig.TLSCaCert != "" {
+		logger.Info("tlsConfig specified, loading tls config")
+		caCert, err := os.ReadFile(tlsConfig.TLSCaCert)
+		if err != nil {
+			return nil, fmt.Errorf("failed to read tls.ca: %w", err)
+		}
+		caCertPool := x509.NewCertPool()
+		caCertPool.AppendCertsFromPEM(caCert)
 
-	// certman watches for newer client certifictes and automatically reloads them
-	cm, err := certman.New(tlsConfig.TLSCert, tlsConfig.TLSKey)
-	cm.Logger(ethLogger{logger: logger})
-	if err != nil {
-		logger.Error("failed to read tls cert or key", "err", err)
-		return nil, err
-	}
-	if err := cm.Watch(); err != nil {
-		logger.Error("failed to start certman watcher", "err", err)
-		return nil, err
-	}
+		// certman watches for newer client certifictes and automatically reloads them
+		cm, err := certman.New(logger, tlsConfig.TLSCert, tlsConfig.TLSKey)
+		if err != nil {
+			logger.Error("failed to read tls cert or key", "err", err)
+			return nil, err
+		}
+		if err := cm.Watch(); err != nil {
+			logger.Error("failed to start certman watcher", "err", err)
+			return nil, err
+		}
 
-	httpClient := &http.Client{
-		Transport: &http.Transport{
-			TLSClientConfig: &tls.Config{
-				MinVersion: tls.VersionTLS13,
-				RootCAs:    caCertPool,
-				GetClientCertificate: func(_ *tls.CertificateRequestInfo) (*tls.Certificate, error) {
-					return cm.GetCertificate(nil)
+		httpClient = &http.Client{
+			Transport: &http.Transport{
+				TLSClientConfig: &tls.Config{
+					MinVersion: tls.VersionTLS13,
+					RootCAs:    caCertPool,
+					GetClientCertificate: func(_ *tls.CertificateRequestInfo) (*tls.Certificate, error) {
+						return cm.GetCertificate(nil)
+					},
 				},
 			},
-		},
+		}
+	} else {
+		logger.Info("no tlsConfig specified, using default http client")
+		httpClient = http.DefaultClient
 	}
+
 	rpcClient, err := rpc.DialOptions(context.Background(), endpoint, rpc.WithHTTPClient(httpClient))
 	if err != nil {
 		return nil, err