+++ packages/contracts-bedrock/tasks/check-l2.ts
@@ -382,7 +382,8 @@ const check = {
 
     await assertSemver(
       OptimismMintableERC20Factory,
-      'OptimismMintableERC20Factory'
+      'OptimismMintableERC20Factory',
+      '1.1.0'
     )
 
     const BRIDGE = await OptimismMintableERC20Factory.BRIDGE()
@@ -547,7 +548,8 @@ const check = {
 
     await assertSemver(
       OptimismMintableERC721Factory,
-      'OptimismMintableERC721Factory'
+      'OptimismMintableERC721Factory',
+      '1.1.0'
     )
 
     const BRIDGE = await OptimismMintableERC721Factory.BRIDGE()